section11/ritual-roast-flask-backend/requirements.txt
Flask==3.1.1
flask-cors==6.0.1
mysql-connector-python==9.3.0
boto3==1.39.3